fix error

+++ dimabe_manufacturing/models/mrp_workorder.py
@@ -246,8 +246,9 @@ class MrpWorkorder(models.Model):
     @api.multi
     def organize_move_line(self):
         for move in self.production_id.move_raw_ids:
-            for active in move.active_move_line_ids:
-                active.unlink()
+            if self.current_quality_check_id.quality_state != 'none':
+                for active in move.active_move_line_ids:
+                    active.unlink()
         for item in self.potential_serial_planned_ids.mapped('stock_production_lot_id'):
             stock_quant = item.get_stock_quant()
             stock_move = self.production_id.move_raw_ids.filtered(lambda a: a.product_id.id == item.product_id.id)
